+++ NEWS.md
@@ -2,6 +2,7 @@ Version 0.8.5-dev
 2012-02-??
 
  - Do not assume any thrown exception is an Error (#65)
+ - Clear both aut and non-auth render cache on style change (#59)
 
 Version 0.8.4
 2013-01-30